Move machine upgrades to a data map
